Required functions written, not tested
The functions are a straight port from the vector example given in the
assignment. makeCacheMatrix generates a list of functions to access the
matrix and its inverse, and to change both quantities.
cacheSolve returns the stored inverse if not null, and otherwise uses
solve to compute the inverse, and stores it.